Frontend implemented

+++ backend/services/ProcessService.py
@@ -1,6 +1,8 @@
 import uuid
 from typing import List, Optional
 from fastapi import Depends, HTTPException
+from models.DocumentationUpdateModel import DocumentationUpdateModel
+from models.AncestorDocumentationModel import AncestorDocumentationModel
 from models.StageModel import StageModel
 from models.ProcessModel import ProcessModel
 from models.DocumentationModel import DocumentationModel
@@ -37,15 +39,17 @@ class ProcessService:
 
         avo_request = AVORequest(**avo.model_dump())
         process.assign_avo_request(avo_request)
+        previous_stage = process.stage
         process.advance_stage()
 
         self.db.add(process.stage)
+        self.db.delete(previous_stage)
         self.db.commit()
         self.db.refresh(process)
 
         return process
 
-    def upload_user_documents(self, process_id: int, user_documents: List[DocumentationModel]) -> ProcessModel:
+    def upload_user_documents(self, process_id: int, user_documents: List[DocumentationUpdateModel]) -> ProcessModel:
         process: Process = self.db.query(Process).filter(Process.id == process_id).first()
         if not process:
             raise HTTPException(status_code=404, detail="Process not found.")
@@ -56,17 +60,19 @@ class ProcessService:
             user_documentation_list.append(document)
 
         process.add_user_documentation(user_documentation_list)
-        selected_documents = [doc for doc in user_documents if doc.file_type.lower() == "pdf"]
+        selected_documents = [doc for doc in user_documents if ".pdf" in doc.name.lower()]
         process.add_attachments_to_translate(selected_documents)
+        previous_stage = process.stage
         process.advance_stage()
 
         self.db.add(process.stage)
+        self.db.delete(previous_stage)
         self.db.commit()
         self.db.refresh(process)
 
         return process
 
-    def upload_avo_documents(self, process_id: int, avo_documents: List[DocumentationModel]) -> ProcessModel:
+    def upload_avo_documents(self, process_id: int, avo_documents: List[DocumentationUpdateModel]) -> ProcessModel:
         process: Process = self.db.query(Process).filter(Process.id == process_id).first()
         if not process:
             raise HTTPException(status_code=404, detail="Process not found.")
@@ -78,35 +84,41 @@ class ProcessService:
 
         process.add_avo_documentation(avo_documentation_list)
         process.add_attachments_to_translate(avo_documentation_list)
+        previous_stage = process.stage
         process.advance_stage()
 
         self.db.add(process.stage)
+        self.db.delete(previous_stage)
         self.db.commit()
         self.db.refresh(process)
 
         return process
 
-    def upload_descendants_documents(self, process_id: int, descendants_documents: List[DocumentationModel]) -> ProcessModel:
+    def upload_ancestors_documents(self, process_id: int, ancestor_documentation: AncestorDocumentationModel) -> ProcessModel:
         process: Process = self.db.query(Process).filter(Process.id == process_id).first()
         if not process:
             raise HTTPException(status_code=404, detail="Process not found.")
         
-        descendant_documentation_list = []
-        for doc in descendants_documents:
-            document = DescendantDocumentation(**doc.model_dump())
-            descendant_documentation_list.append(document)
+        process.ancestor_count = ancestor_documentation.count
+        
+        ancestor_documentation_list = []
+        for doc in ancestor_documentation.documentation:
+            document = AncestorDocumentation(**doc.model_dump())
+            ancestor_documentation_list.append(document)
 
-        process.add_descendant_documentation(descendant_documentation_list)
-        process.add_attachments_to_translate(descendant_documentation_list)
+        process.add_ancestors_documentation(ancestor_documentation_list)
+        process.add_attachments_to_translate(ancestor_documentation_list)
+        previous_stage = process.stage
         process.advance_stage()
 
         self.db.add(process.stage)
+        self.db.delete(previous_stage)
         self.db.commit()
         self.db.refresh(process)
 
         return process
 
-    def upload_translated_documents(self, user_id: int, translated_documents: List[DocumentationModel]) -> ProcessModel:
+    def upload_translated_documents(self, user_id: int, translated_documents: List[DocumentationUpdateModel]) -> ProcessModel:
         user = self.db.query(User).filter(User.id == user_id).first()
         if not user:
             raise HTTPException(status_code=404, detail="User not found.")
@@ -121,9 +133,11 @@ class ProcessService:
             translated_documentation_list.append(document)
 
         process.add_translated_documentation(translated_documentation_list)
+        previous_stage = process.stage
         process.advance_stage()
 
         self.db.add(process.stage)
+        self.db.delete(previous_stage)
         self.db.commit()
         self.db.refresh(process)
 
@@ -134,7 +148,11 @@ class ProcessService:
         if not process:
             raise HTTPException(status_code=404, detail="Process not found.")
 
-        document_list = process.documentations
+        document_list = []
+        document_list.extend(process.user_documentation)
+        document_list.extend(process.avo_documentation)
+        document_list.extend(process.ancestors_documentation)
+        document_list.extend(process.translated_documentation)
 
         return document_list
 
@@ -143,8 +161,8 @@ class ProcessService:
         if not process:
             raise HTTPException(status_code=404, detail="The process to delete does not exist.")
         
-        self.db.delete(process.stage)
-        self.db.delete(process.request_avo)
+        if process.stage: self.db.delete(process.stage)
+        if process.request_avo: self.db.delete(process.request_avo)
         self.db.delete(process)
         self.db.commit()
